Refactor checkpoint interface
